Adjusting decks to show only two regions
Although we have some formats that allow three regions (singleton).
The most player format is standard, which allows only two regions.

With the introduction of Bandle City and multiregion cards, we need to
filter which are the main regions of the deck.

To do this, the code counts how many cards each region have in the decks
and choose the two regions with more cards.